+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Changed
 - In the Files app, use the Music app icon on the file actions menu item "Play"
+- Allow manual entry of file name when exporting a playlist or the radio stations
 
 ### Fixed
 - "New files to scan" and "Scanning" bottom panes being slightly misplaced on NC 25+